+++ internal/action/config.go
@@ -2,13 +2,12 @@ package action
 
 import (
 	"context"
-	"errors"
 	"fmt"
 
 	"github.com/gopasspw/gopass/internal/action/exit"
+	"github.com/gopasspw/gopass/internal/config"
 	"github.com/gopasspw/gopass/internal/out"
 	"github.com/gopasspw/gopass/internal/set"
-	istore "github.com/gopasspw/gopass/internal/store"
 	"github.com/gopasspw/gopass/pkg/ctxutil"
 	"github.com/gopasspw/gopass/pkg/debug"
 	"github.com/urfave/cli/v2"
@@ -34,10 +33,10 @@ func (s *Action) Config(c *cli.Context) error {
 		return exit.Error(exit.Usage, nil, "Usage: %s config key value", s.Name)
 	}
 
-	// need to initialize sub-stores so we can update their configs.
-	// special case: we can always update the root config.
-	if err := s.IsInitialized(c); err != nil && store != "" {
-		return err
+	// sub-stores need to have been initialized so we can update their local configs.
+	// special case: we can always update the global config. NB: IsInitialized initializes the store if nil.
+	if inited, err := s.Store.IsInitialized(ctx); err != nil || (store != "" && !inited) {
+		return exit.Error(exit.Unknown, err, "Store %s seems uninitialized or cannot be initialized", store)
 	}
 
 	if err := s.setConfigValue(ctx, store, c.Args().Get(0), c.Args().Get(1)); err != nil {
@@ -80,8 +79,17 @@ func contains(haystack []string, needle string) bool {
 func (s *Action) setConfigValue(ctx context.Context, store, key, value string) error {
 	debug.Log("setting %s to %s for %q", key, value, store)
 
-	if err := s.cfg.Set(store, key, value); err != nil {
+	level, err := s.cfg.SetWithLevel(store, key, value)
+	if err != nil {
 		return fmt.Errorf("failed to set config value %q: %w", key, err)
+	}
+
+	// in case of a non-local config change we don't need to track changes in the store
+	if level != config.Local && level != config.Worktree {
+		debug.Log("did not set local config (was config level %d) in store %q, skipping commit phase", level, store)
+		s.printConfigValues(ctx, store, key)
+
+		return nil
 	}
 
 	st := s.Store.Storage(ctx, store)
@@ -89,12 +97,25 @@ func (s *Action) setConfigValue(ctx context.Context, store, key, value string) e
 		return fmt.Errorf("storage not available")
 	}
 
-	if err := st.Add(ctx, "config"); err != nil && !errors.Is(err, istore.ErrGitNotInit) {
-		return fmt.Errorf("failed to stage config file: %w", err)
+	configFile := "config"
+	if level == config.Worktree {
+		configFile = "config.worktree"
 	}
-	if err := st.Commit(ctx, "Update config"); err != nil && !errors.Is(err, istore.ErrGitNotInit) {
-		return fmt.Errorf("failed to commit config: %w", err)
+
+	// notice that the cfg.Set above should have created the local config file.
+	if !st.Exists(ctx, configFile) {
+		return fmt.Errorf("local config file %s didn't exist in store, this is unexpected", configFile)
 	}
+
+	if err := st.TryAdd(ctx, configFile); err != nil {
+		return fmt.Errorf("failed to Add local config %q: %w", configFile, err)
+	}
+	debug.Log("Added local config for commit")
+
+	if err := st.TryCommit(ctx, "Update config"); err != nil {
+		return fmt.Errorf("failed to commit local config: %w", err)
+	}
+	debug.Log("Committed local config")
 
 	s.printConfigValues(ctx, store, key)