+++ internal/action/otp.go
@@ -8,6 +8,7 @@ import (
 	"time"
 
 	"github.com/gopasspw/gopass/internal/action/exit"
+	"github.com/gopasspw/gopass/internal/config"
 	"github.com/gopasspw/gopass/internal/out"
 	"github.com/gopasspw/gopass/internal/store"
 	"github.com/gopasspw/gopass/pkg/clipboard"
@@ -110,6 +111,8 @@ func (s *Action) otp(ctx context.Context, name, qrf string, clip, pw, recurse bo
 		return s.otpHandleError(ctx, name, qrf, clip, pw, recurse, err)
 	}
 
+	outerCtx := ctx
+	ctx = config.WithMount(ctx, s.Store.MountPoint(name))
 	ctx, cancel := context.WithCancel(ctx)
 	defer cancel()
 
@@ -162,8 +165,9 @@ func (s *Action) otp(ctx context.Context, name, qrf string, clip, pw, recurse bo
 			}
 			counter++
 			_ = sec.Set("counter", strconv.Itoa(int(counter)))
-			if err := s.Store.Set(ctx, name, sec); err != nil {
-				out.Errorf(ctx, "Failed to persist counter value: %s", err)
+			// using outerCtx here because we want to save the counter even if the user cancels.
+			if err := s.Store.Set(outerCtx, name, sec); err != nil {
+				out.Errorf(outerCtx, "Failed to persist counter value: %s", err)
 			}
 			debug.Log("Saved counter as %d", counter)
 		}